Refactor/SearchModal: 파일명 변경(SearchHeader > SearchModalHeader)
