+++ src/components/SearchModal/SearchModalHeader.module.scss
